refactor(bridge): extract incoming action classifier
